Moved simple types into folders
#organization
