+++ include/archipelago_common.py
