+++ views/view_confirm_force_disconnect.py
